// src/generated/server/alsaglobal/infrastructure/v1/service_server.ts
export interface ListInternetOutagesRequest {
  start: number;
  end: number;
  pageSize: number;
  cursor: string;
  country: string;
}

export interface ListInternetOutagesResponse {
  outages: InternetOutage[];
  pagination?: PaginationResponse;
}

export interface InternetOutage {
  id: string;
  title: string;
  link: string;
  description: string;
  detectedAt: number;
  country: string;
  region: string;
  location?: GeoCoordinates;
  severity: OutageSeverity;
  categories: string[];
  cause: string;
  outageType: string;
  endedAt: number;
}

export interface GeoCoordinates {
  latitude: number;
  longitude: number;
}

export interface PaginationResponse {
  nextCursor: string;
  totalCount: number;
}
// ...
export type OutageSeverity = "OUTAGE_SEVERITY_UNSPECIFIED" | "OUTAGE_SEVERITY_PARTIAL" | "OUTAGE_SEVERITY_MAJOR" | "OUTAGE_SEVERITY_TOTAL";
// ...
export interface ServerContext {
  request: Request;
  pathParams: Record<string, string>;
  headers: Record<string, string>;
}